requirements.txt
fastapi
uvicorn
pydantic-settings
sqlalchemy
asyncpg
alembic
pgvector
aiogram
openai
faster-whisper
sentence-transformers
